Start work on splitting up library into modules
